Added centos check detection
Added centos 7 check detection

+++ functions/check_linux.sh
@@ -3,14 +3,14 @@
 #    Check Installed Linux Version
 #---------------------------------------------------------------------
 
-apt-get -y install lsb-release
-
 CheckLinux() {
+if [ -f /etc/debian_version ]; then
+  apt-get -y install lsb-release
   echo -n "Checking your installed Linux Version... "
  
-#---------------------------------------------------------------------
-#    Debian 7 Wheezy
-#---------------------------------------------------------------------
+  #---------------------------------------------------------------------
+  #    Debian 7 Wheezy
+  #---------------------------------------------------------------------
   
   if command -v lsb_release &> /dev/null; then
 	if lsb_release -a 2> /dev/null | grep -iq "wheezy"; then
@@ -18,9 +18,9 @@ CheckLinux() {
 	fi
   fi
 
-#---------------------------------------------------------------------
-#    Debian 8 Jessie
-#---------------------------------------------------------------------
+  #---------------------------------------------------------------------
+  #    Debian 8 Jessie
+  #---------------------------------------------------------------------
   
   if command -v lsb_release &> /dev/null; then
 	if lsb_release -a 2> /dev/null | grep -iq "jessie"; then
@@ -28,15 +28,22 @@ CheckLinux() {
 	fi
   fi
 
-#---------------------------------------------------------------------
-#    Ubuntu
-#---------------------------------------------------------------------
+  #---------------------------------------------------------------------
+  #    Ubuntu
+  #---------------------------------------------------------------------
   
   if command -v lsb_release &> /dev/null; then
 	if lsb_release -a 2> /dev/null | grep -iq "ubuntu"; then
 		DISTRO=ubuntu
 	fi
   fi
+fi
+
+if [ -f /etc/centos-release ]; then
+  if [ `cat /etc/centos-release | grep 7.0 | wc -l` -ne 0 ]; then
+        DISTRO=centos7
+  fi
+fi
 
   # ONLY for Debug..... :)
   echo -e "Your Distro is: " $DISTRO
@@ -46,5 +53,6 @@ CheckLinux() {
   then
     exit 1
   fi
+
 }
